+++ src/main/scala/package.scala
@@ -11,7 +11,7 @@ import com.github.tototoshi.csv.*
 
 /** Chromatin fiber tracing with FISH probes */
 package object looptrace {
-    val VersionName = "0.3.1"
+    val VersionName = "0.3.2"
 
     type CsvRow = Map[String, String]
     type ErrorMessages = NonEmptyList[String]